+++ mirage/factories/meeting-submission.ts
@@ -2,7 +2,11 @@ import { Factory, faker } from 'ember-cli-mirage';
 
 import MeetingSubmission from 'ember-osf-web/models/meeting-submission';
 
-export default Factory.extend<MeetingSubmission>({
+export interface MirageMeetingSubmission extends MeetingSubmission {
+    created: Date | string;
+}
+
+export default Factory.extend<MirageMeetingSubmission>({
     title() {
         return faker.lorem.sentence();
     },
@@ -18,13 +22,19 @@ export default Factory.extend<MeetingSubmission>({
     downloadCount() {
         return faker.random.number({ min: 5, max: 500 });
     },
-    created() {
+    dateCreated() {
         return faker.date.past(5, new Date(2020, 0, 0));
     },
 });
 
+declare module 'ember-cli-mirage/types/registries/model' {
+    export default interface MirageModelRegistry {
+        'meeting-submission': MirageMeetingSubmission;
+    } // eslint-disable-line semi
+}
+
 declare module 'ember-cli-mirage/types/registries/schema' {
     export default interface MirageSchemaRegistry {
-        meetingSubmissions: MeetingSubmission;
+        meetingSubmissions: MirageMeetingSubmission;
     } // eslint-disable-line semi
 }